Added: New passive scalars, which now work properly and can be disabled with a preprocessor directive (-DSCALARS). Parts of a mechanism to set variables manually in HOC are now implented, but they are not quite ``production quality'' yet. Dependencies in the Makefile are somewhat better now, though not necessarily perfect.  I discovered that the LH section only works on GNU make--need to fix that at some point.
Modified:
After much hair-pulling and searching, it was discovered that rain was using
a uninitialized value in the top of the VNr z-level, which is now set to 0.
The subroutine tridag now has attributes to prevent unintentional changes in
the arguments.
The use of optional arguments required the change of the some subroutines
into modules. Particularly closure_new, gcss, sounding, mixing, and diag_var.
I cleaned up polpak.f90 a little more, adding "end subroutine *" to some of
the routines.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@104 f1ce5500-ea2b-45e0-a110-28dffb978352