Two changes: It was noted that the tridiagonals were not quite right at the corners of the matrix--Dr. Larson helped make a new boundary condition to compensate.
I put rain in its own module.  This will protect the scoping of all the helper
functions, because they are now private, and make all the parameters strongly
typed to avoid errors when they're passed (e.g. it will check to make sure that
a variable isn't a 32 bit real when being passed as a 64 bit DP).


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@508 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 61a1a2ee0f37c7696c91cef89daee41e2c8443b9