Fix for an old bug.
After 90 minutes, clear air radiation and advection are supposed to be added
into the thlm_forcing variable for ATEX, but when I added bugsrad I didn't
declare thlm_forcing for cloud_rad() as in/out and it was not being added in
anymore.  This shouldn't affect FIRE, because thlm_forcing is always 0.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@626 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: a8fb8262b355c399c5b0b4565df8bb1179ca75e0