Committed modifications to hoc.F and statistics.F to fix statistics sampling problem.  Deleted unused variables 'ioutput' and 'spunup'.
The statistics GrADS files from HOC now follow the same averaging
convention as the COAMPS stats files.  The time in the GrADS file corresponds
to the end of the averaging period.  For example, if

stats_tsamp =  60.0
stats_tout  = 300.0

the first time in the GrADS file will be 00:05 and the fields for that
time will correspond to the average taken over the first 5 minutes of the
simulation and sampled every 60 seconds.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@639 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: f1c13670d3b12b50396aa405cd6a31418265d5bb