Committed modifications to diag_var.F to fix variances and covariance budgets. These variables are diagnostic and hence their budgets balance in a different way compared to prognostic variables:
The sum of the _ta _tpx _dp1 _dp2 _ma terms should be zero (except at
the lowest model level due to surface boundary conditions not taken into
account in the budgets). The term _cl reflects the effect of the clipping
(i.e. difference in the field after and before the clipping is applied).
The term _bt is the change in the variable compared to the previous
time step. Both _cl and _bt terms are not part of the diagnostic budget.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@641 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: b637762fd82093a4ef84195b6f6cfef26776131a