Changes:
Added files outputnc and outputfile;  The former contains subroutines for
outputing to netcdf format, the latter defintions for both netcdf and GrADS
output.

Changed the name of the last subroutine is statistics.F to stats_finalize,
since now it calls a subroutine to finish writing a netcdf file as well as
deallocates the arrays.

Changed stats_zero and stats_average to take non derived type arguments.
The advantage of this is that now stats_zero only takes intent(in) and
intent(out) arguments and should be easier for a compiler to vectorize.
My hope was that I could do the same for stats_average, but this was not
the case.

Added the units field to statistics.F and the outputfile_class module;  While
GrADSnc is not configured to take advantage of this for anything but time,
there are apparently other udunits enable programs that can.  Most of the
variables are set to units "unknown" right now, since there are a lot of
variables and those I was uncertain about I didn't want to guess on.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@654 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: d8ed6cd6e36847031dd4b22eb1aec301e8517118