2 Major changes to the code: Added: gamma_coefb, and gamma_coefc parameters.  Gamma can now be a function of Skw.
Modified:
C2 is no longer split into rt, thl, and rtthl terms.  Now it has a b and c
term and is a function of Skw as well.

Broken:
The budget terms tuner for terms that are function of C2 or gamma.  If this
change works out, we'll take the time to fix it, but it's a lot of code.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@726 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 4d7b15f825e0b8cc16ed30f413157bc42e79eb4e