Cleaned up some comments.
Added the constant parameter fstderr so error messages aren't redirected
with the stdout stuff.

Changed the instances when a solution is singular within working precision
to set isValid to .true., where before it wasn't being set at all.

Changed the shifting within band_solve and band_solvex.  I had problems
figuring out what my own code does, and hopefully it is now more clear.
It may have had a bug for cases where nsub was not equal to nsup before
as well, but I'm not sure.

Changed the band_solve routine so that it copies over the LHS and then
shifts it, rather than shifting the LHS matrix and then copying.  The LHS
array can therefore be intent(in) now, and this should make the subroutine
easier to optimize since the LULHS array is a local variable.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@1040 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 23a0eaf2d403ce67ea6946b7d30b06f056c210e6