I changed the file rain.F so that the correlations and variance/mean^2 parameters are set at every height level instead of as a constant parameter.  This does two things:
1)  It sets up the proper infrastructure in the code so that we can easily
    implement the actual equations for correlations, if we ever get the
    proper variables to do so.
2)  It sets up the proper infrastructure in the code so that we can
    implement two different sets of correlations, one for in-cloud and
    one for below-cloud (just as in our Griffin-Larson rain
    parameterization notes).

For now, I have set both the in-cloud and below-cloud parameters equal to
each other (their values being what was in the code before).  Therefore,
it should produce identical results to what we had before as long as the
code was set up right.  I checked this out, and indeed we got identical
results to what we had before.  Therefore, everything is fine with the
new code.

The next step is to try different parameter sets for in-cloud and
below-cloud and see how that affects the results!


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@1107 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: bf08aa7d15553d53bd9f79caf7082111cab8776e