I changed the files that had something to do with rain.  I changed the value for the minimum rcm to officially have a cloud from the magic number 1.0e-6 to the variable rc_tol.  The variable rc_tol can be found in constants.F.  For these files, the value of rc_tol affects which levels we set Ncm.
The value of rc_tol also affects statistics.F when searching for cloud
base.  Again, the value of "minimum cloud" was changed from the magic
number 1.0e-6 to rc_tol.  In addition to that, the file rain.F was
affected for the same reasons (in determining which constants are used,
in-cloud or below-cloud).

For rain.F, I set both Ncp2_Ncm2_below and corr_sNc_NL_below to 0 and
wrote a comment about how these variables are not applicable below cloud.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@1142 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 60d37eca77a95c4cdd41f868ee72fe6e69247574