Changed case specific Nov11 and Jun25 settings for a grid that matches the actual height in meters for Nov11 and Jun25.
Changed length to account for this so lmin will decrease with
altitude up to 500m (and thus be zero for June 25th and November 11th)

Changed the ldown cutoff for the bug that Prof. Larson found. (min vs. max)

Eliminated bottom_at_sfc flag in favor of specifying the lowest point on the
momentum grid (namelists must be updated accordingly).
All references to bottom_at_sfc now look at gr%zt(1).

Changed the sounding to allow for a lowest point on z that is not near surface.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@1281 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: eeba04e9845fdd31f563edcebe1f1146b26d3868