Added hydrometeor eddy diffusivity constant c_Krrm to the code.
Added the constant c_Krrm to the code to be used for solving for the
eddy diffusivity of hydrometeors.  The equation used is:
Kr = c_Krrm * Kht,
where c_Krrm is a constant and I have tuned it to be 0.075.
The constant nu_r remains in use as background eddy diffusivity.  It's
value right now is 3.
The overall diffusion constant for the hydrometeor code is Kr + nu_r.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@1411 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: cde66149c6086270e18a8edef36b4e8afdcdbed3