ARM 0003 and ARM 97 had grid sizes increased, LBA had coamps microphysics set to true, and all three cases had Khm_aniso set to true.  ARM 0003 had BUGSrad temporarily set to false.
diag_var.F was modified to clip up2 and vp2 at zero.  A clipping statistic
was also added for both up2 and vp2, so hoc_stats.F and stats_init_zm.F
also needed to be modified.



The standalone.in files for ARM 0003, ARM 97, and LBA were altered so that
all three cases now use the exact same set of constants that all other
base cases use.



all_stats.in was modified to add up2_cl and vp2_cl.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@1703 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 417fb5b99241053dc25fb25ee47b2ea756b339b2