An updated sounding to the Nov.11 case.  After some code changes that occurred between March 29 and April 1, the previous Nov.11 sounding produced inexplicably different results.  Therefore, it was necessary to update the sounding to re-match the CLUBB initial profiles with the LES results.
This new sounding is very similar to the original Nov.11 sounding, with the
exception of the additional sounding point needed to extend the model's
vertical domain.  A few small adjustments were made to rt and theta to allow
the t=61 profiles to match LES conditions.

NOTE: This update also required a modification to the "em" profile in hoc.F.
These updates further improved the t=61 results, and also removed some of the
excess jaggedness that appeared in the rcm, rtm and thlm profiles.  See hoc.F
for more details.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@1907 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 7d0bfafcd7278b5f0643c71aabde71e904f992d1