Adding code to save "snowslope" and "Nsnowm" statistics variables.  These variables are useful when examining behaviors from COAMPS microphysics.
NOTE: In order to obtain reasonable results for these variables, I needed
to apply a correction when "snowslope" has very small values (e.g. < 2.0).
If "snowslope" is small, then "Nsnowm" has values larger than 2e7.  The
correction does not need to be applied to COAMPS, so it may be necessary
to update the correction in the future.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@1972 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 3833da918b8d02e27f3e825db1932399e16efa62