I added many descriptive comments to the diffusion module and functions.
I changed a few internal variable names to include an underscore "_".

I caught a mistake in the signs for diffusion_mlev for the uppermost and
lowermost levels.  I corrected that mistake.  Fortunately, none of
the momentum level variables that have eddy diffusion terms have zero-flux
boundary conditions.  They all have fixed-point boundary conditions.  That
means that the uppermost and lowermost levels are never called in
diffusion_mlev.

I ran the new code and got an exact binary match for all cases and files.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@1978 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 3dc328a16b5c02e1a46584ec62f726094fa70075