I have changed the RF02 grid files to contain true stretched grids.
The RF02 zm_grid files are taken directly from source code on the RF02
website.  The RF02 intercomparison specifies a stretched grid that is
based on 97 momentum levels covering a 1500 m. vertical domain.  The
thermodynamic levels are placed halfway between the momentum levels.
This can be used in CLUBB by selecting grid_type = 3.

The RF02 zt_files are based on the stretched grid that is found in SAM.
The stretched grid found in SAM uses the same source code from the
RF02 website as mentioned above, but alters it so that it takes
148 momentum levels to cover the 1500 m. vertical domain.  The source
code places the thermodynamic levels halfway between the momentum levels.
However, in the case of SAM, the thermodynamic levels are output from
the source code.  SAM reads the thermodynamic levels as input, and then
places it's momentum levels halfway between the thermodynamic levels.
Thus, the resulting momentum levels in SAM are at slightly different
altitudes than the momentum levels output by the original source code.
This can also be used in CLUBB by selecting grid_type = 2.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@2138 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: c75a79f548c82d5433e53b973bcf0cd126242cbc