I made the following changes to diag_var.F:
1)  I removed wp2 from the call list to diag_var_lhs and diag_var_rhs
because it isn't being used in either of those subroutines anymore (wp2_zt
is being used instead).

2)  In subroutine diag_var_uv_rhs, I found that the statistical code that
was there had the terms dp1 and pr1 switched around.  In other words, it
was outputting dp1 as pr1 (and pr1 as dp1) in the stats budgets.  I
corrected this bug.

3)  I added a few small comments and cleaned up the code in general.

A run_standalone-all.bash (which doesn't produce the statistical budget
terms) found that there was no binary difference in any GrADS .dat file
for any case.  I did a run_standalone.bash (which calls all_stats.in by
default) for the BOMEX case in order to see the results of the stats
budget change.  I found that I had indeed corrected the mistake that was
there.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@2191 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: c997d633f08335114b57b08581c8835769dcd289