I have discovered that the Jun.25 SCM case was previously set up so that its initial profiles match LES results at t=1 min.  However, we want those profiles to match at t=61 min.  Therefore, I have updated the temperature, moisture and wind profiles so that the profiles match at t=61.
The profiles are not perfect, but they should be an improvement over
previous results.  This sounding produces a bit too much liquid in both
cloud layers, but the qtm, thlm, um and vm profiles are all very close
to LES results.  Cloud fraction in the partly cloudy layer is a bit too
high (~80%), but it is still better than the 100% we got in the previous
version.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@2216 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 050916b2ca78e48a2b517fccc0f2859be5391f2d