The code now clips wprtp, wpthlp, and wpsclrp at three points:
1) after diag_var.F, where rtp2, thlp2, and sclrp2 are set;
2) after mixing.F, where wprtp, wpthlp, and wpsclrp2 are set; and
3) after wp23.F, where wp2 is set.

Any change to those values changes the correlation of w and rt,
the correlation of w and thl, or the correlation of w and sclr.

In addition, the semi-implicit clipping code is now commented into
mixing.F by default.

A plotgen shows no change to almost all of the cases.  Only MPACE-A has
been significantly altered.  However, the cloud fields still remain
basically the same, and the rest of the fields still look good.

Furthermore, this code allows ARM 97 and ARM 0003 to run to completion
(with c_Ksqd = 0) with the least amount of instability found yet.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@2396 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: ff7a54f3d96138a73122be52e58cc9496e3f225e