I changed the code to enforce the zero-flux boundary condition for um/vm/edsclrm at the upper boundary.  I reflected that change in stats. A plotgen showed no difference for most cases; and only a slight difference near the upper boundary in a couple cases.  The contour results for the ARM deep convective cases also remain the same.
I got rid of the code that uses a linear extension to solve for um and vm
at the upper and lower boundary.

Now, um and vm at the lowest (below ground) level are set in relation to
the surface flux, such that:
xm(1) = xm(2) + xpwp_sfc / ( Kh_zm(1) * gr%dzm(1) ).  This does not effect
the rest of the code, as um(1) or vm(1) is not currently referenced.
However, it keeps the value of um(1) or vm(1) consistent with the surface
flux and the value of um(2) or vm(2).


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@2730 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: e48474a5e784f8d5aa72488307a6a1fcb87504bb