I changed wm_zt_p1 back to wm_ztp1 and wm_zm_m1 back to wm_zmm1.  The reason for doing so is that the '_p1' and '_m1' formats are not consistent with the formats used in the rest of the model.  For example, when wp2 is sent into a function on three levels, it is sent is as wp2p1, wp2, and wp2m1; not wp2_p1, wp2, and wp2_m1.  Likewise, when wp3 is interpolated to momentum levels (as wp3_zm) and sent into a function on two levels, it is sent in as wp3_zmp1 and wp3_zmm1; not wp3_zm_p1 and wp3_zm_m1.
As made clear by the above examples, the format throughout the model has 
been to tack on 'p1' and 'm1' to the end of the variable name, without an
underscore '_'.  All variable naming conventions should follow the same format.
If the format is changed, it should be changed throughout the model.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@2792 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 42fca60ee7599906b1d96e3159cbeea572533148