I altered microphys_driver.F90 so that the lhs array loops over levels 2...gr%nnzp, and then computes the boundary levels outside of the loop.  This was done to keep the code consistent with the code found in diag_var.F90, mixing.F90, wp23.F90, and compute_um_edsclrm.F90, which all obey the same coding structure.
A plotgen revealed no difference in any cases, with the exception of some slight 
differences in MPACE-A.  This can be chalked up to the fact that the new code 
is not exactly a binary match to the old one.  In the old version, the boundaries 
were part of the loop, and therefore time tendency was computed first, and then 
eddy-diffusion.  However, now at the boundaries, eddy-diffusion is computed first, 
followed by time tendency.  Since MPACE-A reacts to even the slightest binary change,
this comes as no suprise.  All other microphysics cases show everything to be the same.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@2793 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 6bd770da0975fc07472c7da7054d9e0b3594ebee