Corrected an error in the non-expert routines, where err == ndim + 1 was interpreted to mean the matrix was singular within working precision, which is only true for the expert routines.
Changed tridag_solvex and band_solvex to emit some warnings at debug_level = 2
when the forward error estimate is greater than 3 digits.  Usually this only
happens the model is unstable, but it is one more diagnostic to determine if
something is going wrong with the direct solvers.

Ran findent on the source code.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@2948 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 98366d932584fcc1f2aaeb90f287fce8f99b0cd2