I added new code for use in stretched grid cases.  The values of the constant background coefficients of eddy diffusivity (nu1, nu2, nu6, nu8, nu9, and nu_r) are now based on either the average value of the grid spacing (for stretched grid cases) or the value of deltaz (traditional method; for evenly-spaced grid cases).  For the way that CLUBB is set up right now, the only cases that are effected by this change are LBA, ARM 97, and ARM 0003.
As a additional note, SAM-CLUBB has been running by sending in the value
of height at momentum level 1 (SAM's zi(1)) as the input for CLUBB's "deltaz"
(this can be found in SAM's call to "setup_clubb_core").  Since the height of
SAM's momentum level 1 always set to 0 (found in SAM's setgrid.f90), a value
of 0 has been sent into CLUBB setup as "deltaz" all this time.  The CLUBB code
for adjusting the values of nu# is based on having a deltaz greater than 40 m.
(or 20 m. for nu_r).  Thus, SAM-CLUBB cases with grid spacings of greater than
40 m. were not having the values of nu# adjusted proportionally with grid spacing,
even though their counterparts in CLUBB standalone were being adjusted properly.
I have corrected this problem by placing code in CLUBB that adjusts the values of
nu# by using the average grid spacing if CLUBB is implemented in a host model.

Finally, I also added code to adjust the value of the constant coefficient for
4th-order hyper-diffusion, nu_hd, based on either the average value of the
grid spacing (for stretched grid cases and cases where CLUBB is implemented
inside a host model) or the value of deltaz (for evenly-spaced grid cases).
However, unlike the rest of the nu# coefficients, where the value of nu# is
only adjusted if the grid spacing is greater than 40 m. (or 20 m. for nu_r),
the value of nu_hd is only un-adjusted if the average grid spacing is exactly
40 m.  If the grid spacing is smaller than 40 m., nu_hd is adjusted to be smaller,
proportionally.  If the grid spacing is larger than 40 m., nu_hd is adjusted to be
larger, proportionally.  This is done because the value of nu_hd has much stronger
effects based on grid spacing than the nu-terms relating to 2nd-order diffusion.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3001 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 1fc1afda1ac07f1ab2240ec109e5d55e86c785af