A corrected a bug in compare_plots_cases.m.
When a surface variable is not found in a given GrADS sfc file,
it's value is supposed to be set to zero for every element in
the array.  However, the previous code contained the lines:

avg_prev_values(1:nz_prev) = 0.0;

and

avg_curr_values(1:nz_curr) = 0.0;

which set the value of the variable to 0 for every index in a
VERTICAL array.  Since this is a surface variable, a vertical
array is not needed.  Rather, a time array is needed.  I replaced
the first line above with:

avg_prev_values(t1_prev:t2_prev) = 0.0;

and the second line above with:

avg_curr_values(t1_curr:t2_curr) = 0.0;

Now, plotgen works correctly again.




git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3082 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: e7d4e546ed620d24e3390a8259494355f91a2696