I modified run_standalone_all.bash in regards to the '--timestep_test' option.  I have commented out the previous code, which set dtmain, dtclosure, stats_tsamp, and stats_tout to the declared time step length.  I have placed a comment above it which states:
# Use this version if statistical output is desired.

The user can un-comment that code (and comment out the
new code) if statistical output is desired.

Immediately below that code, I have written some new
code that sets dtmain and dtclosure to the declared
time step length and sets l_stats to .false.  I have
placed a comment above the new code which states:

# Use this version is statistical output is not desired.

The new version is currently in use.  The code is now
set up so that run_timestep_tests.bash can be used as
part of the nightly tests.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3083 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 20e9f01be96836b1d7842f3ccb2e49a3cb775168