I added code to damp u'^2 and v'^2 to a minimum threshold value of wtol_sqd (rather than to 0), just like the code is currently set up for w'^2.  The other variances in the code are also currently damped to a threshold value.
The damping term for u'^2 and v'^2 (pressure term 1) was:

 - (2/3) * ( C_14 / tau_zm ) * em.

It is now:

 - (2/3) * ( C_14 / tau_zm ) * ( em - emin );

where emin = (3/2) * wtol^2.

This works out so that both u'^2 and v'^2 are damped to a value
of wtol_sqd.  The value of wtol_sqd is also the hole-filling and
clipping threshold for both u'^2 and v'^2.

This was done because:

1)  it keeps the code for variances u'^2 and v'^2 consistent
    with the code for the other variances in the model (w'^2,
    r_t'^2, th_l'^2, and sclr'^2); and

2)  I have found through past experiments that higher values
    of TKE (em) generally lead to increased model stability.
    This is due to the facts that (a) a greater value of em
    leads to a smaller value of tau_zm/tau_zt, and thus a
    larger damping factor; and (b) a greater value of em leads
    to a greater value of Kh_zm/Kh_zt, and thus a larger amount
    of eddy diffusivity.

The results don't effect any of the base cases much at all.
The time-height contours of the constantly shifting ARM 97 have
increased to +/- 1000 m^3/s^3 for w'^3 and 0 to 130 m^2/s^2 for
w'^2.  I'm sure that these will come right back down again with
my next SVN update.

I performed a run_standalone_all.bash for standard (1 minute),
5 minute, 10 minute, 15 minute, 20 minute, 30 minute,
45 minute, and 60 minute time steps.  The results are as follows:

standard (1 minute):  All cases run to completion.
5 minute:   All cases run to completion.
10 minute:  All cases run to completion.
15 minute:  All cases run to completion.
20 minute:  Only RICO blows up.
30 minute:  Only RICO blows up.
45 minute:  Only RICO blows up.
60 minute:  RICO and ARM 97 blow up.

The results are good, although most of the progress was made by
the application of hole-filling to rtm/thlm on Friday.




git-svn-id: file:///usr/local/src/svnroot/clubb_repos/hoc_v2.2_tuner/trunk@3096 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: d900153342ae4930bb589e4d82ac3c90d03ac63d