I have been working on revising the model in regards to when error messages are printed.  The goal is to have only error messages associated with fatal errors printed to the error output stream when debug level is set to 0.  All error outputs not associated with fatal errors should only print when the debug level is at least 1 (this includes warning messages).
All clubb_at_least_debug_level, print, and write statements have
been thoroughly checked throughout the model code.  The remaining
adjustments that needed to be made are now being committed.  The
adjustments made within the Latin_hypercube and BUGSrad directories
have not been tested, although they compile and they should work
properly.  Every other change in this code has been extensively
tested.

The next step is to work on the run_scm_all.bash script so that
standard output is suppressed, but standard error is allowed to
print to screen.  However, a 'sed' command will have to be used
to change debug_level to 0, so that only fatal error messages are
printed to the screen during a run_scm_all.bash.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3267 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: b981cbd308407b44a7e4e2a7a1890e47cb81d220