I have changed inputfields to allow for LES data to be read in from a LES using a different grid than the grid being used for the CLUBB run. The new code tests whether a given CLUBB level is at exactly the same altitude as an LES level, whether the given CLUBB level falls between two LES levels, or whether the given CLUBB level is outside of the domain of the LES data.  If the given CLUBB level falls between two LES levels, a linear interpolation of the LES data is performed for the given CLUBB level.
I have maintained the structure of the previous code as much as
possible in order to maintain the results for a "standard" case, where
the LES levels and the CLUBB levels exactly match.  For case ( "les" ),
I have been able to exactly match prior inputfields results for FIRE
(which is the only case that I tested).  For case ( "rf1" ), which is a
special case for the DYCOMS-II RF01 LES data, I have been able to match
prior inputfields results to within numerical error.

This new inputfields code is very versatile.  It allows for special tests
like running the BOMEX case with a very coarse, deep-convective-case-style,
stretched grid while using LES input at the lower levels.

The inputfields code has not yet been modified for cases where the input
is from CLUBB data.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3305 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 1e77cbdbfa6f60a7bf383b5721d086eda611cd62