This change was made because of the fact that too much diffusivity was occurring at very coarse grid spacings.  This was due to the fact that the values of the background eddy diffusivity coefficients were being increased by a linear factor when the average vertical grid spacing exceeded 40 meters.  The linear factor was (avg_delta_z/40.0).  Thus, a vertical grid that averaged 200 m. spacing had its values of background diffusivity coefficients increased by a factor of 5, and a vertical grid that averaged 1000 m. spacing had its values of background diffusivity coefficients increased by a factor of 25.  Thus, simulations that used very coarse grid spacings (for example, 1000 m.) were being over-diffused.
The solution:

The first part of the solution was to turn off the code that adjusted the
values of the background eddy diffusivities, and concentrate on adjusting
the coefficient of variable diffusivities (c_K1 ... c_K9).  The goal was to
find a solution that could still nearly match results to the standard set of
CLUBB runs, while maintaining smooth profiles for "standard resolution" ARM 97 
and LBA runs (both which use a vertical grid that averages around 213 meters
of vertical spacing).  From that point, it should be much easier to find a
good agreement between a case run at low resolution and a case run at high
resolution.

At standard resolutions, once the background diffusivity adjustment code
was turned off, the profiles for ARM 97 and LBA became very, very noisy.
The change in parameters in tunable_parameters.in helped to combat this noise
while providing good results for the rest of the benchmark cases.  However,
this by itself was not enough, as "standard resolution" ARM 97 and LBA were
still too noisy for comfort.

The next part of the solution was to re-activate the code that adjusted the
values of background eddy diffusivities, but to change the code so that the
increase is not as drastic.  At first, a code with a multiplier of the form
SQRT( avg_delta_z / 40.0 ) was used, which would multiply the background
diffusivity coefficients by SQRT(5) at 200 m. grid spacing and by SQRT(25) (=5)
at 1000 m. grid spacing.  However, this code still turned out to be not
diffusive enough for "standard resolution" ARM 97 and LBA cases.

Thus, a new multiplier was used, which had the form:

   1.0 + 2.0 * LN( avg_delta_z / 40.0 ).

Thus, at exactly 40 meter average vertical grid spacing, the value of the
multiplier is 1, and thus the values of the eddy diffusivities remain as listed
in tunable_parameters.in.  This allows for continuity, as the values of the
background eddy diffusivities are not to be altered at average grid spacings
at or below 40 meters.  However, above 40 meters, the multipler increases --
sharply at first -- but levels off as average grid spacing gets larger.

Thus, at 50 m. grid spacing (only 10 m. above the multiplying threshold), the
value of the multiplier is approximately 1.44; at 200 m. grid spacing, the value
of the multiplier is approximately 4.2; but at 1000 m. grid spacing, the value of
the multiplier is only approximately 7.4.  This allows for the "standard resolution"
ARM 97 and LBA runs (that appear on the nightly plots) to be sufficiently smoothed
out, but it limits the amount of diffusivity at very coarse grid spacings.  This
formula can be tinkered with a bit as needed.





git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3461 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 05b8c55eeffd062acad710ad6f34d3d588c602ae