I made a small adjustment to the multiplication factor for background eddy diffusivities (which is activated when the average vertical grid spacing exceeds 40 meters and l_adj_low_res_nu is turned on).
The formula was changed from:

   1.0 + 2.0 * LN( avg_delta_z / 40.0 )

to:

   1.0 + 1.5 * LN( avg_delta_z / 40.0 ).

The change in the value of the coefficient in front of the LN( ) term
is small but significant.  While the "old" formulation improved results
for very coarse resolution BOMEX, the newest formulation is great
improvement even upon those results.  Yet, the newest formulation still
keeps the profile results smooth for ARM 97 and LBA at their "standard"
resolutions.  Finally, the benchmark cases still look good.

Now, at 50 m. grid spacing (only 10 m. above the multiplying threshold),
the value of the multiplier is approximately 1.33; at 200 m. grid spacing,
the value of the multiplier is approximately 3.41; but at 1000 m. grid
spacing, the value of the multiplier is only approximately 5.83.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3464 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: fff73aef86953123c6fbcb8ec4867cbf018753a2