These changes are all related to ticket #41
In the matrix_operations module I removed the numerical recipes gaussj 
subroutine and matrix_mult.  In place of these I added a wrapper to the
LAPACK subroutine for LU decomposition and used the matmul intrinsic in the
lh_sampler subroutines.  This makes no difference in the results for BOMEX
or DYCOMS II RF02 DS.

In the lh_sampler subroutine I modified the subroutines to adhere to coding
standards, though I'm not sure I caught everything.

I also ran findent on both these files.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3485 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 1e910cc64d318ae3cb991bafea2aab7c7d226f2b