Changed height calculations in the vertical_avg function in fill_holes.
This should be more accurate (in terms of less potential for round-off errors), and faster to execute.

A warning: this commit will likely change the outputs of the ARM 97 and TWP_ICE cases--this is likely due to the combination of deep convection, microphysics, and potentially the type of stretched grid used.

This commit addresses ticket #59.

~EIHoppe//20090702


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3717 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: a3355b9853a373aae5a8a0870180e34807c464af