I altered the code so that the cloud water turbulent flux, w'r_c' (variable name wprcp), is output from advance_clubb_core.  This was done so that w'r_c' could be provided to WRF-CLUBB.  This change won't effect the standalone CLUBB results.  I didn't add wprcp to the list of variables that need to be read into CLUBB for a restart run because wprcp is not a necessary input into CLUBB.  I have already altered SAM-CLUBB to account for this change to the advance_clubb_core call list.
While I was at it, I organized the variable declarations in subroutine
advance_clubb_core so that the reason each variable is being passed into
or out of advance_clubb_core is clear.  For the input/output variables,
there are four sections:  (1) prognostic variables or variables that
are closely related to the prognostic variables (this includes
subsections for passive scalars and eddy passive scalars; (2) variables
that need to be saved from timestep-to-timestep because they are used
in advance_clubb_core before they are reset in advance_clubb_core;
(3) variables that are output for use in other parts of CLUBB, such as
microphysics, forcings, and BUGSrad; and (4) variables that are output
for use in a host model.




git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3889 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 7639cc171facf30d0702351c8243872bf79a543b